functions/pagination/requirements.txt
google-cloud-storage
google-cloud-pubsub
google-cloud-error_reporting
google-cloud-logging
bs4
lxml
unidecode

